+++ action.yml
@@ -42,38 +42,37 @@ runs:
     - name: Set HEAD_REF
       shell: bash
       run: echo "HEAD_REF=${{github.event.pull_request.head.ref || github.ref_name}}" >>"${GITHUB_ENV}"
+    - name: Set env
+      id: set-env
+      shell: bash
+      run: |
+        echo "PR_DESCRIPTION_PREFIX=${{inputs.pr-description-prefix}}" >> "$GITHUB_ENV"
+        echo "PR_NUMBER=${{github.event.pull_request.number}}" >> "$GITHUB_ENV"
+        echo "PR_TITLE_PREFIX=${{inputs.pr-title-prefix}}" >> "$GITHUB_ENV"
+        echo "BRANCH_NAME_PREFIX=${{inputs.branch-name-prefix}}" >> "$GITHUB_ENV"
     # 差分があったときは、コミットを作りpushする
     - name: Push
       env:
-        PR_TITLE_PREFIX: ${{inputs.pr-title-prefix}}
         TOKEN: ${{inputs.github-token}}
-        BRANCH_NAME_PREFIX: ${{inputs.branch-name-prefix}}
         NO_VERIFY: ${{inputs.no-verify}}
       if: steps.diff.outputs.result != '' && ((github.event_name == 'pull_request' && github.event.action != 'closed') || github.event_name == 'push' || github.event_name == 'schedule' || github.event_name == 'workflow_dispatch' || github.event_name == 'repository_dispatch')
       working-directory: ${{inputs.working-directory}}
       run: ${{ github.action_path }}/scripts/action/push.sh
       shell: bash
-    - name: Get PullRequests
+    - name: Get a number of PullRequests
       uses: actions/github-script@v7.0.1
       if: steps.diff.outputs.result != '' && ((github.event_name == 'pull_request' && github.event.action != 'closed') || github.event_name == 'push' || github.event_name == 'schedule' || github.event_name == 'workflow_dispatch' || github.event_name == 'repository_dispatch')
-      id: get_pull_requests
-      env:
-        BRANCH_NAME_PREFIX: ${{inputs.branch-name-prefix}}
+      id: get_number_of_pull_requests
       with:
         github-token: ${{inputs.github-token}}
         script: |
-          const script = require('${{ github.action_path }}/scripts/action/get_pull_requests.js')
+          const script = require('${{ github.action_path }}/scripts/action/get_number_of_pull_requests.js')
           return await script({github, context})
     # pushしたブランチで修正PRを作る
     - name: Create PullRequest
       uses: actions/github-script@v7.0.1
-      if: steps.diff.outputs.result != '' && steps.get_pull_requests.outputs.result == 0 && ((github.event_name == 'pull_request' && github.event.action != 'closed') || github.event_name == 'push' || github.event_name == 'schedule' || github.event_name == 'workflow_dispatch' || github.event_name == 'repository_dispatch')
+      if: steps.diff.outputs.result != '' && steps.get_number_of_pull_requests.outputs.result == 0 && ((github.event_name == 'pull_request' && github.event.action != 'closed') || github.event_name == 'push' || github.event_name == 'schedule' || github.event_name == 'workflow_dispatch' || github.event_name == 'repository_dispatch')
       id: create_pull_request
-      env:
-        PR_DESCRIPTION_PREFIX: ${{inputs.pr-description-prefix}}
-        PR_NUMBER: ${{github.event.pull_request.number}}
-        PR_TITLE_PREFIX: ${{inputs.pr-title-prefix}}
-        BRANCH_NAME_PREFIX: ${{inputs.branch-name-prefix}}
       with:
         github-token: ${{inputs.github-token}}
         script: |
@@ -82,7 +81,7 @@ runs:
     # 元のPRを出したユーザーを修正PRにアサインする
     - name: Assign a user
       uses: actions/github-script@v7.0.1
-      if: steps.diff.outputs.result != '' && steps.get_pull_requests.outputs.result == 0 && github.event_name == 'pull_request' && github.event.action != 'closed' && github.event.pull_request.user.login != 'dependabot[bot]' && github.event.pull_request.user.login != 'renovate[bot]'
+      if: steps.diff.outputs.result != '' && steps.get_number_of_pull_requests.outputs.result == 0 && github.event_name == 'pull_request' && github.event.action != 'closed' && github.event.pull_request.user.login != 'dependabot[bot]' && github.event.pull_request.user.login != 'renovate[bot]'
       env:
         PR_NUMBER: ${{steps.create_pull_request.outputs.result}}
       with:
@@ -90,12 +89,20 @@ runs:
         script: |
           const script = require('${{ github.action_path }}/scripts/action/assign_a_user.js')
           await script({github, context})
+    # 修正PRのタイトルやDescriptionを更新する
+    - name: Update PullRequest
+      uses: actions/github-script@v7.0.1
+      if: steps.diff.outputs.result != '' && 0 < steps.get_number_of_pull_requests.outputs.result && ((github.event_name == 'pull_request' && github.event.action != 'closed') || github.event_name == 'push' || github.event_name == 'schedule' || github.event_name == 'workflow_dispatch' || github.event_name == 'repository_dispatch')
+      id: update_pull_request
+      with:
+        github-token: ${{inputs.github-token}}
+        script: |
+          const script = require('${{ github.action_path }}/scripts/action/update_pull_request.js')
+          return await script({github, context})
     # 既に修正PRがある状態で、手動でformatを修正した場合、修正PRを閉じる
     - name: Close PullRequest
       uses: actions/github-script@v7.0.1
       if: (github.event_name == 'pull_request' && (github.event.action == 'closed' || steps.diff.outputs.result == '')) || ((github.event_name == 'push' || github.event_name == 'schedule' || github.event_name == 'workflow_dispatch' || github.event_name == 'repository_dispatch') && steps.diff.outputs.result == '')
-      env:
-        BRANCH_NAME_PREFIX: ${{inputs.branch-name-prefix}}
       with:
         github-token: ${{inputs.github-token}}
         script: |
